pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "tempcast"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.8",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "tensorflow>=2.21.0",
]

[project.scripts]
tempcast = "tempcast:main"
assets = "tempcast.core.cli:assets"

[dependency-groups]
dev = ["ipykernel>=7.2.0", "ruff>=0.15.11", "wandb>=0.26.0"]
